requirements.txt
langchain
langchain-openai
langchain-tavily
python-dotenv
openai
jupyter
wikipedia